requirements.txt
scikit-learn==1.5.2
xgboost==2.1.1
nltk==3.9.1
pandas==2.2.3
numpy==2.1.3
dvc==3.51.2
PyYAML==6.0.2
dagshub==0.3.27
mlflow==2.16.2
s3fs==2025.2.0
setuptools<81
httpx==0.23.3
jupyterlab<4.0